major changes

+++ Bookings/handleCancelBooking.js
@@ -2,6 +2,7 @@ import { getCollection } from "../utility/readDB.js";
 import fs from "fs/promises";
 import path from "path";
 import { fileURLToPath } from "url";
+import { isBookingExpired } from "../utility/checkBookingExpiry.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -24,21 +25,27 @@ export const handleCancelBooking = async (req, res) => {
       throw new Error("No booking found with the provided code");
     }
 
-    //delete associated image if any
+    if (isBookingExpired(existingBooking)) {
+      res.writeHead(404, { "Content-Type": "application/json" });
+      return res.end(JSON.stringify({ error: "Cannot cancel past bookings" }));
+    }
+
+    // Delete associated images if any
     if (existingBooking.imagePaths && existingBooking.imagePaths.length > 0) {
       for (const imagePath of existingBooking.imagePaths) {
         const absolutePath = path.join(sampleImages, imagePath);
-
         try {
           await fs.unlink(absolutePath);
         } catch (error) {
-          console.warn(`filed to delete image: ${absolutePath}`, error);
+          console.warn(`Failed to delete image: ${absolutePath}`, error);
         }
       }
     }
 
-    // Remove booking
-    await bookingsCollection.deleteOne({ id: bookingId });
+    await bookingsCollection.updateOne(
+      { id: bookingId },
+      { $set: { status: "cancelled" } }
+    );
 
     res.writeHead(200, { "Content-Type": "application/json" });
     res.end(JSON.stringify({ message: "Booking canceled successfully" }));
